[core] Fix #7: decreasing fixpoint iterations do not converge
In some cases, the decreasing iterations of the fixpoint computation
would not converge. This was due to 2 errors in the implementation of
the forward fixpoint iterator:

* The narrowing operator assures the convergence only if it is applied
right after a widening. The current implementation was using a meet at
the first step, thus breaking the convergence assurance.

* The convergence check was done before applying the narrowing. This is
supposed to be done afterwards.

This patch also adds a test that triggered a divergence.