[analyzer] Fix #14: disable sqlite journal and synchronization
On Windows, the initialization of the output database is very slow.

Since we never need to rollback, we disable the database journal using
`PRAGMA journal_mode = OFF`.

Since we are not concerned about the integrity of the database if the OS
crashes, we disable the full synchronization using `PRAGMA synchronous =
OFF`.

These changes make the database faster on Windows.