[llvm-to-ar] Improve the design of the type translation
I created 3 different helper classes:
* `TypeWithSignImporter`: translate a type with a preferred signedness
* `TypeWithDebugInfoImporter`: translate a type with a debug info
* `TypeMatcher`: check if two types match

The `TypeWithDebugInfoImporter` throws a `TypeDebugInfoMismatch` if a
mismatch between llvm type and debug info occurred. This is better than
checking with a `match()` functions, since on success there is no need to
re-traverse the type. It also avoids code duplication (`match_di_type()` was a
copy/paste of `translate_di_type()`) and bugs such as #44 and #46.

Internally, `TypeWithDebugInfoImporter` can fork to create a child
`TypeWithDebugInfoImporter` and try to translate a type. If a mismatch
occurs, the child is discarded. Otherwise, the internal mapping can be
saved. This is used when translating a union type, for instance.