[core][analyzer] Add a narrowing with threshold operator
The narrowing with threshold operator should be used to regain precision
after reaching a fixpoint using a widening with threshold.

Using the regular narrowing after a widening with threshold could lead to
a lose of precision because the usual narrowing only shrinks infinite bounds.

For instance, with intervals and a threshold of 10:
Increasing iterations:
`[0, 1] -> [0, 2] -> [0, 3] -> widening w/ threshold -> [0, 10]`
Decreasing iterations with usual narrowing:
`[0, 10] -> [0, 9] -> narrowing -> [0, 10]`
Decreasing iterations with narrowing with threshold of 10:
`[0, 10] -> [0, 9] -> narrowing -> [0, 9]`