[cmake] Build and run a small program to find the versions of the dependencies
The previous code parses the header files to find the version of GMP,
MPFR, PLL and SQLite. This was not robust (e.g, see #143). This patch
compiles and runs a small program that prints the version of each
dependencies. This is more robust and also makes sure that the libraries
actually work.