[analyzer] Remove the "level of precision" command line option
This option is mostly unused and complicates the numerical execution
engine implementation. This patch removes it.

This patch also adds the `ExecutionEngineOptions` to replace the `Precision`
parameter. It allows to control whether the numerical execution engine should
maintain the allocation size variable or not. It is set ON for the value
analysis and OFF for the pointer analysis.