Detect dynamic uninitialized variables with debug info optional
1. Fix a problem in the tracking of the values of
dynamically-allocated memory so that use of
uninitialized dynamically-allocated memory is
detected.

2. Allow analysis of programs without debug information.
This is important for the case where the LLVM
comes from a binary lifter where no debug information
may be available. Without debug information the name
of the function with a problem is reported but
file and line number information is not available.