[analyzer] Handle formatting path in drive different from cwd (#165).
The report-view command throws an exception when executed in Windows
from a path in a drive different from that of the files being analyzed.

This is due to the use of os.path.relpath on two paths in potentially
different drives as part of the definition of format_path. According to
relpath's definition, the function throws a ValueError on Windows if the
two arguments provided are on different drives.

This commit modifies format_path so that it first checks if the paths
are on different drives, returning the absolute path in that case, and
executing the old behavior otherwise.