Make git ignore build and install as directories, but not as files (#217).
The .gitignore rules are matching build and install filenames and directories
outside of the root of the tree, which causes any file in the project with
those names to be ignored.

This commit modifies the .gitignore to match only directories with the names
build and install, not files. This is standard in projects that use cmake.