[analyzer] Open server on IPv6 when available (#158).
This commit makes ikos-scan and ikos-view use IPv6 when available. More
specifically, it replaces uses of HTTPServer, which by default only
listens on IPv4 interfaces, with calls to a custom class HTTPServerIPv6.
The latter listens on IPv6 when available, as well as IPv4 whenever
dual-stack is supported. If IPv6 is not available, the server defaults
to the old behavior implemented by HTTPServer.