[analyzer] Skip llvm-objdump lines by format, not count (#203).
IKOS uses `llvm-objdump` and expects the header of the `.bc` file to have 4
lines. However, in llvm14, that header has 3 lines instead. This results in an
error message:

/usr/local/llvm14/bin/llvm-link: No such file or directory

This commit modifies the processing of the output of `llvm-objdump` to be based
on the content, skipping all lines until it finds one that matches the format it
expects.