Remove Travis CI script, badge (#186).
This repo still has a .travis.yml, but Travis CI is no longer running on the
repo. The badge on IKOS' README file shows that the build status is unknown.

The Travis CI build script is also outdated (it uses LLVM9) and would no longer
run. Although it is not being built by Travis CI when we push to the IKOS
repo, Travis will still attempt to run it on forks if the owners of such forks
have Travis CI credits. The build then fails, giving the impression that the
tool is broken.

This commit removes the Travis CI compilation script altogether, as well
as the badge in the README file.