[analyzer] Use pip to install package (#233).
The current installation scripts for python programs don't place files exactly
where/how expected, resulting in errors like:

```
$ ikos
error: could not find ikos python module
error: see TROUBLESHOOTING.md
```

even when PYTHONPATH is set correctly.

Instead of trying to fix the issue, it's worth it to switch to an alternative
that is more up-to-date, where the problem does not happen.

The current installation scripts for python programs included in IKOS use
easy_install, which has been deprecated
(https://setuptools.pypa.io/en/latest/history.html#v58-3-0).

This commit replaces calls to the setup.py script, with calls to pip, which is
one of the recommended solutions for python package installation.