[analyzer] Install the python module inside a virtual environment (#243).
The current installation process tries to install the ikos python module
as a system library. This is probably not what a typical user want.

Instead, let's install a virtual environment for ikos and install the
python module within it. Following homebrew's standard, this is
installed under <install-path>/libexec.

Note that this also deprecates Python 2 since virtual environments
require Python 3.3