Install IKOS in HOME directory in CI job. Refs #275.
The CI job for Mac seems to have updated the image and it's failing to
build when it used to work before. The permissions of top-level
directories in the image now prevent us from installing ikos at the top
level.

This change makes the CI job install IKOS in the user's HOME directory
instead.