[frontend] Add translation for LLVM fneg instruction (#197).
IKOS does not understand the LLVM unary instruction fneg, which returns the
negation of its operand. This makes common C programs fail to be processed by
IKOS, which stops with a message like:

```
ikos-analyzer: <some_source.pp.bc>: error: unsupported llvm instruction fneg [2]
ikos: error: a run-time error occurred
```

This commit adds to IKOS' frontend the translation of the fneg instruction.  It
also adds two tests that use fneg, one with float and one with double, which
should be analyzable with IKOS.