[analyzer] Remove use of deprecated python module pipes (#292).
The analyzer currently uses pipes in older versions of python, and shlex
only when it's available. A prior commit to IKOS required Python 3.3,
where shlex was introduced, so it is safe to assume that it's available
unconditionally. Additionally, the module pipes was deprecated in python
3.11 and finally removed in python 3.13. This is making ikos not work
with modern versions of Python.

This commit replaces uses of pipes' quote method with shlex's quote.