[frontend] Replace calls to deprecated function getAlignment (#304).
The function llvm::AllocaInst::getAlignment was deprecated in llvm-14
and removed in llvm-15. To facilitate IKOS working with newer versions
of llvm, we should replace calls to getAlignment().

This commit replaces several occurrences of getAlignment with calls to
getAlign().value(). The replacement is only applied in cases where it is
clear that the an alignment exists (the function getAlign returns an
Align), as opposed to places where the alignment may be optional (where
the function getAlign returns a MaybeAlign).